HTTP phase 2: give the TCP/IP stack its own file, so HTTP can raise it
netman + smsutils + smap + ps2ip + ps2ips + the HTTP RPC client are one shared
resource -- one SMAP NIC, one driver, same cost whoever asked -- but they lived in
ethsupport.c and could only be reached through the SMB session code. HTTP has to
raise the NIC without implying a share. src/netsupport.c now owns the stack;
ethsupport.c keeps the session, the shares, the credentials and SMB's error strings.

The move is verbatim. I checked that rather than asserting it: every relocated body
diffs identical to its original under the eth->net rename, the only textual change
being two log strings that no longer say "SMB" about a stack that is no longer only
SMB's.

netDeinitModules is the one function whose shape changed, and it is the one worth
reading. The protocol's own teardown is a callback so the sequence is preserved by
construction instead of by argument. I got that wrong first: hoisting SMB's two flag
writes into a hook that ran before NetManDeinit moved them earlier than they had ever
run. Safer, probably -- a racing ethIsSMBShareConnected() would stop seeing a live
share sooner -- but "probably safer" is an unevidenced behaviour change in teardown
code nobody can test on hardware today, which is the whole reason the deviation rule
exists. nbnsDeinit now stays in netsupport next to HttpDeinit (both are RPC clients
the stack registered), which frees the hook to sit exactly where the flag writes were.
The teardown diff is now those two assignments becoming a call, and nothing else.

NIC exclusivity falls out of the sharing and must stay that way. Every claimant tests
netGetModulesLoaded(), and netLoadModules() refuses when UDPBD or the UDPFS chain
already drives SMAP -- so because HTTP raises the very same stack, all three existing
interlocks already cover it and no fourth check is added anywhere. The header says so,
because the tempting mistake later is to give HTTP a private modules-loaded flag, and
that is exactly what would put two drivers on the EMAC.

netEnsureModules() exists because SMB alone wants the stack up WITHOUT applying the
network configuration -- it applies it later inside its own connect sequence. Anything
else wants netLoadInitModules().

ethGetModulesLoaded / ethLoadInitModules / ethDeinitModules / ethApplyConfig /
ethGetNetConfig / ethGetDHCPStatus stay as delegations, so not one of the seventeen
call sites across gui, opl, bdmsupport, udpfssupport, debug, ranet and vcdsupport
moves in this commit.

Builds clean (MAKE_EXIT=0, ps2dev:latest, from make clean) and clang-format 12 clean.
No hardware test -- nothing here is observable without one, which is why the plan
flags this phase for review on its own before HTTP exists to blame.
